LayoutTests/platform/gtk/imported/w3c/web-platform-tests/mathml/presentation-markup/operators/mo-minsize-maxsize-001-expected.txt


FAIL minsize < 0 is treated as 0 assert_approx_equals: expected 125 +/- 1 but got 25
PASS maxsize < minsize is treated as maxsize = minsize
PASS minsize < maxsize < 0 is treated as maxsize = minsize = 0
PASS target size = 0 is treated as Tascent = Tdescent = minsize/2
FAIL minsize/maxsize percentages are relative to the target size assert_approx_equals: percent minsize expected 300 +/- 1 but got 25
FAIL default minsize/maxsize percentages assert_approx_equals: default maxsize is infinity expected 7500 +/- 1 but got 25
⥯

⥯

⥯

⥯

⥯

⥯

⥯

⥯
